refactor: 旧UI Playmat経路を撤去し現役dialogs/ruleActionCandidatesをgame/へ移設(判定者監査済)
本番デフォルトで到達不能だった旧UI(dev fixture ?ui=legacy 専用)をユーザー授権で撤去(D4前倒し)。

- 削除: Playmat.tsx + 旧UI12ファイル(Battlefield/GameLog/Hand/InfoPanel/MobileControlsDrawer/
  MobileZoneSwap/PlaymatHud/Stack/TargetPickerDialog/Toasts/TriggerCandidatePanel/Zones)+
  各テスト + RotateNotice.tsx。playmat/ ディレクトリは消滅。
- 撤去: App.tsx の VITE_UI_V2_LAYOUT ロールバック分岐と isV2LayoutEnabled()、fixture ?ui=legacy。
  既定は常に GameScreen(挙動不変=旧分岐は既定 true で到達不能だった)。
- 移設(git mv・相対import不変): 現役の dialogs.tsx / ruleActionCandidates.ts と各テストを
  src/components/game/ へ。新UI(actionCatalog/gameController)が常時 import する現役コード。
- import追随: actionCatalog / gameController / review.m64 / ZoneViewerLimit / OpponentSetupScreen。
- 死CSS: 直接結合の .playmat-shell / .rotate-notice* / portrait media を除去。旧UI由来の
  大区画~580行(.playmat*/.hand__*/.stack__*/.mobile-*/.battlefield*)の purge は共有クラスの
  個別到達性解析を要するため別スライスへ分離(docs/ui-architecture-v2.md §6 で追跡)。
- check:forbidden 修正: FORBIDDEN regex を /(^|\/|\.)review\./ にし Name.review.test.tsx も検出
  (CardPreview.tsx は誤検出せず)。
- docs正本更新: README(削除しないガードを授権解除へ)/ ui-architecture-v2 §4·§5·§6·D4行 /
  acceptance memo / design-playbook / visual-fixtures README / gameController ヘッダ /
  台帳 d4c-legacy-removal(部分実行を追記・status=deferred 維持)。

独立Tier-1監査(冷Sonnet・2巡)PASS。npm run check 全緑(lint/2019 tests/build)。
実機(fixture battlefield/stack)GameScreen 正常描画・コンソール0件・?ui=legacy はGameScreenへ無害化。
