+++ docs/acceptance.md
@@ -1,7 +1,7 @@
 # G1/G2 受け入れシナリオ(E2E)
 
-> **現行性メモ(2026-07-12・J0草稿)**: 本文には旧Playmat由来の右クリック／DnDシナリオと
-> 新GameScreenの経路が混在する。PC回復(D4a)では、どちらのUIを対象にする期待値かを分離し、
+> **現行性メモ(2026-07-19 更新)**: 旧 Playmat UI は削除済み(2026-07-19)ゆえ、本文の
+> 右クリック／DnDシナリオは**すべて新 GameScreen を対象に読む**。PC回復(D4a)では、
 > 既定GameScreenでDnD・hover・keyboard・zone操作が実行できるまで「PC同等性green」と扱わない。
 > 判定者再オーナー化要。
