+++ docs/design-playbook.md
@@ -8,6 +8,11 @@
 > ユーザー退行が確認された。本書の既存D4カードは「PC回復＋3カラム＋旧実装大量削除」を結合しており、
 > 現状のまま実行しない。D6/D7も停止する。回復案=`research/design/design-recovery-plan.draft.md`。
 > D4分割は判定者の格上げ事項(judge-protocol §7)、優先順位はユーザー、不在期の変更は復帰判定者の再オーナー化が必要。
+>
+> **更新(2026-07-19・ユーザー授権)**: 上記結合カードのうち**「旧実装大量削除」だけを単独で実行**した
+> (旧 Playmat + 周辺12ファイル削除・`dialogs.tsx`/`ruleActionCandidates.ts` を `game/` へ移設)。
+> 既定経路が本番デフォルトで到達不能=生きた同等性参照ではなかったための撤去で、**PC回復＋3カラムは
+> 依然停止**(別途 D4a で実施)。詳細 = `docs/README.md`・`docs/ui-architecture-v2.md` §6 負債。
 
 ---
