+++ research/design/visual-fixtures/README.md
@@ -7,8 +7,7 @@
 Vite起動後:
 
 ```text
-http://127.0.0.1:5173/research/design/visual-fixtures/?scenario=hand&ui=new
-http://127.0.0.1:5173/research/design/visual-fixtures/?scenario=hand&ui=legacy
+http://127.0.0.1:5173/research/design/visual-fixtures/?scenario=hand
 ```
 
 `scenario`:
@@ -20,7 +19,7 @@ http://127.0.0.1:5173/research/design/visual-fixtures/?scenario=hand&ui=legacy
 - `stack` — battlefield＋stack 2＋warnings
 - `graveyard` — stack scene＋graveyard 10＋exile 2
 
-`ui`は`new`または`legacy`。同じbuilder出力を使うため、新旧比較でshuffleや操作履歴がずれない。
+レンダ先は新レイアウト`GameScreen`のみ(旧`Playmat`とその`?ui=legacy`経路は 2026-07-19 に削除)。
 
 このページは視覚比較器であり、受け入れ判定そのものではない。D4a reviewと5 viewport実機確認が
 別途必要。
