+++ src/App.tsx
@@ -1,8 +1,6 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import './App.css';
 import { ImportScreen } from './components/ImportScreen';
-import { RotateNotice } from './components/RotateNotice';
-import { Playmat } from './components/playmat/Playmat';
 import { GameScreen } from './components/game/GameScreen';
 import { OpponentSetupScreen } from './components/game/OpponentSetupScreen';
 import { Modal } from './components/Modal';
@@ -25,15 +23,6 @@ import type { CardDef } from './types/card';
 
 const DECK_TEXT_KEY = 'mtg-onedeck:deck-text';
 const DECK_CARDS_KEY = 'mtg-onedeck:deck-cards';
-
-/**
- * ゲーム画面の描画先。既定=新レイアウト GameScreen(D2・縦持ち第一級)。
- * VITE_UI_V2_LAYOUT=false で旧 Playmat+RotateNotice へ即時ロールバック
- * (docs/ui-architecture-v2.md §4 strangler)。関数化して描画時に env を読む(テスト両値可)。
- */
-function isV2LayoutEnabled(): boolean {
-  return import.meta.env.VITE_UI_V2_LAYOUT !== 'false';
-}
 
 interface StoredDeckCard {
   def: CardDef;
@@ -229,31 +218,12 @@ function App() {
         />
       );
     }
-    if (isV2LayoutEnabled()) {
-      // D2: 縦持ち第一級の新レイアウト(RotateNotice なし=縦持ちの壁を撤去)。
-      return (
-        <GameScreen
-          keybindings={keybindings}
-          onOpenOpponentSetup={() => setGameView('opponent-setup')}
-        />
-      );
-    }
-    // ロールバック経路: 旧 Playmat(横/デスクトップ既存挙動)+ RotateNotice。
+    // D2: 縦持ち第一級の新レイアウト(縦持ちの壁を撤去)。
     return (
-      <div className="playmat-shell">
-        <button
-          type="button"
-          className="btn"
-          onClick={() => setGameView('opponent-setup')}
-          data-testid="legacy-open-opponent-setup"
-        >
-          対戦相手セットアップ
-        </button>
-        <div className="playmat-shell__game">
-          <Playmat keybindings={keybindings} />
-        </div>
-        <RotateNotice />
-      </div>
+      <GameScreen
+        keybindings={keybindings}
+        onOpenOpponentSetup={() => setGameView('opponent-setup')}
+      />
     );
   }
