+++ src/components/game/ruleActionCandidates.ts
